+++ changelog.md
@@ -24,6 +24,10 @@
 * no longer allow rendering Hibiki HTML directly to 'body' tag (bad interactions with 3rd party libraries)
 * 'foreach' will now skip keys that start with '@' on ojects.  to iterate over them, use fn:objallkeys().
 * added fn:objkeys(), fn:objatkeys() and fn:objallkeys() to return object keys (also work with HibikiWrappedObj).  objkeys omits any keys starting with '@', objallkeys returns all keys (include @), and objatkeys only returns @ keys.
+* added 'type' and 'data' getters to HibikiError object
+* http module errors set error.type to 'http' and error.data to an object with 'status', 'statustext', and 'data' (parsed error response).  can be used to inspect and use error responses (e.g. JSON error bodies with non-200 status codes)
+* if http module encounters unparsable json, it will throw an error, and set error.data.data to a blob with mimetype 'text/json-unparseable'
+* http network errors get error.data.status set to 599, and err.data.statustext to 'Network Error'
 * bugfix: inconsistencies in accessing getters on HibikiNode object
 * bugfix: more consistent handling of noattr args in jsfuncs (stripped out by HibikiParamsObj)
 * bugfix: rendering of text inside of html option tag